Extra organizer add component
