+++ backend/controllers/auth.ts
@@ -3,6 +3,7 @@ import { getNtnuiToken, getNtnuiProfile, refreshNtnuiToken } from "ntnui-tools";
 import { User } from "../models/user";
 import { GroupType } from "../types/user";
 import { groupOrganizers } from "../utils/user";
+import { Organizer } from "../models/organizer";
 
 export async function login(req: Request, res: Response) {
   try {
@@ -45,17 +46,27 @@ export async function login(req: Request, res: Response) {
         organizer: mainAssemblyOrganizer,
       },
     ];
-    userProfile.data.memberships.forEach((membership) => {
+    for (const membership of userProfile.data.memberships) {
       let organizer = false;
+      // User is automatically an organizer if they are part of the group board
       if (groupOrganizers().includes(membership.type)) {
         organizer = true;
+      } else {
+        // Check if user is added as an extra organizer for the given assembly
+        const extraOrganizer = await Organizer.exists({
+          ntnui_no: userProfile.data.ntnui_no,
+          assembly_id: membership.slug,
+        });
+        if (extraOrganizer) {
+          organizer = true;
+        }
       }
       groups.push({
         groupName: membership.group,
         groupSlug: membership.slug,
         organizer: organizer,
       });
-    });
+    }
 
     // Create or update user
     await User.findByIdAndUpdate(
