+++ frontend/src/components/EditAssembly.tsx
@@ -3,15 +3,10 @@ import {
   Box,
   Button,
   Container,
-  Divider,
   Flex,
-  Group,
   Image,
-  Input,
   Loader,
   Modal,
-  Select,
-  SimpleGrid,
   Text,
 } from "@mantine/core";
 import { useEffect, useState } from "react";
@@ -20,21 +15,18 @@ import Arrow from "../assets/Arrow.svg";
 import { UserDataGroupType } from "../types/user";
 import {
   activateAssembly,
-  addExternalOrganizerToAssembly,
   deleteAssembly,
   getAssemblyByName,
-  getExternalAssemblyOrganizers,
-  removeExternalOrganizerFromAssembly,
-  searchForGroupMember,
 } from "../services/assembly";
 import { getVotations } from "../services/votation";
-import { AssemblyType, ExtraOrganizerType } from "../types/assembly";
+import { AssemblyType } from "../types/assembly";
 import VotationPanel from "./VotationPanel";
 import { OptionType, VoteType } from "../types/votes";
 import { Results } from "./Results";
 import { useMediaQuery } from "@mantine/hooks";
 import useWebSocket from "react-use-websocket";
 import { showNotification } from "@mantine/notifications";
+import { AddOrganizerButtonModal } from "./AddOrganizer";
 
 export function EditAssembly(state: { group: UserDataGroupType }) {
   const breakpoint = useMediaQuery("(min-width: 1010px)");
@@ -45,17 +37,8 @@ export function EditAssembly(state: { group: UserDataGroupType }) {
   const [submittedVotes, setSubmittedVotes] = useState<number>(0);
   const [isChanged, setIsChanged] = useState(false);
   const [openModal, setOpenModal] = useState(false);
-  const [openAddOrganizerModal, setOpenAddOrganizerModal] = useState(false);
-  const [ntnui_no_of_new_organizer, setNtnui_no_of_new_organizer] =
-    useState<number>(0);
   const [statusChanges, setStatusChanges] = useState(assembly?.isActive);
   const [accordionActiveTabs, setAccordionActiveTabs] = useState<string[]>([]);
-  const [extraOrganizers, setExtraOrganizers] = useState<ExtraOrganizerType[]>(
-    []
-  );
-  const [searchResult, setSearchResult] = useState<
-    { first_name: string; last_name: string; ntnui_no: number }[]
-  >([]);
 
   const { lastMessage, sendJsonMessage } = useWebSocket(
     import.meta.env.VITE_SOCKET_URL + "/organizer",
@@ -97,19 +80,7 @@ export function EditAssembly(state: { group: UserDataGroupType }) {
     };
 
     fetch().catch(console.error);
-    fetchOrganizers();
   }, [statusChanges]);
-
-  const fetchOrganizers = async () => {
-    console.log("Fetching organizers");
-    const organizers = await getExternalAssemblyOrganizers(group.groupSlug);
-    setExtraOrganizers(organizers);
-  };
-
-  const searchGroupMember = async (search: string) => {
-    const result = await searchForGroupMember(group.groupSlug, search);
-    setSearchResult(result);
-  };
 
   // Triggered when a votation is created, edited, deleted (started/stopped).
   useEffect(() => {
@@ -328,125 +299,7 @@ export function EditAssembly(state: { group: UserDataGroupType }) {
             )}
 
             <br />
-            <Button onClick={() => setOpenAddOrganizerModal(true)} m={10}>
-              <Modal
-                opened={openAddOrganizerModal}
-                onClose={() => setOpenAddOrganizerModal(false)}
-                size="lg"
-                centered
-                transition="fade"
-                transitionDuration={200}
-                exitTransitionDuration={200}
-                // Styling is done like this to overwrite Mantine styling, therefore global color variables is not used.
-                styles={{
-                  modal: {
-                    backgroundColor: "#1b202c",
-                    color: "white",
-                    border: ".5px solid",
-                    borderRadius: 5,
-                    borderBottomRightRadius: 0,
-                    borderColor: "#f8f082",
-                  },
-                  title: {
-                    margin: "0 auto",
-                  },
-                }}
-              >
-                <Text mb={5} fw={600} ta={"center"}>
-                  Add organizer
-                </Text>
-                <Text ta={"center"}>
-                  The user need to be a member of the group.
-                </Text>
-
-                <Flex direction="row" align="center" justify="center">
-                  <Box
-                    m={20}
-                    sx={(theme) => ({
-                      border: `1px solid ${theme.colors.gray[3]}`,
-                      borderRadius: theme.radius.md,
-                      padding: theme.spacing.md,
-                      minWidth: 300,
-                    })}
-                  >
-                    {extraOrganizers.length < 1 ? (
-                      <Text align="center">No extra organizers</Text>
-                    ) : (
-                      <Text mb={20} align="center">
-                        Extra organizers:
-                      </Text>
-                    )}
-                    {extraOrganizers.map((organizer, index) => (
-                      <Box key={organizer.ntnui_no}>
-                        <Group position="center" noWrap grow={true}>
-                          <Text sx={{ flex: 1 }}>{organizer.name}</Text>
-                          <Button
-                            maw={100}
-                            onClick={() => {
-                              removeExternalOrganizerFromAssembly(
-                                group.groupSlug,
-                                organizer.ntnui_no
-                              ).then(() => fetchOrganizers());
-                            }}
-                            data-testid="remove-organizer-button"
-                          >
-                            Remove
-                          </Button>
-                        </Group>
-                        {index < extraOrganizers.length - 1 && (
-                          <Divider mt={5} mb={5} />
-                        )}
-                      </Box>
-                    ))}
-                  </Box>
-                </Flex>
-
-                <Select
-                  label="Search for group member"
-                  placeholder="Type to search"
-                  searchable
-                  clearable
-                  data={
-                    searchResult.map((member) => ({
-                      value: member.ntnui_no.toString(),
-                      label: `${member.first_name} ${member.last_name}`,
-                    })) || []
-                  }
-                  onSearchChange={(value) => {
-                    searchGroupMember(value);
-                  }}
-                  onChange={(value) => {
-                    if (!value) return;
-                    const selectedMember = searchResult.find(
-                      (member) => member.ntnui_no.toString() === value
-                    );
-
-                    if (selectedMember) {
-                      addExternalOrganizerToAssembly(
-                        group.groupSlug,
-                        selectedMember.ntnui_no,
-                        `${selectedMember.first_name} ${selectedMember.last_name}`
-                      ).then(fetchOrganizers);
-                    }
-                  }}
-                />
-
-                <Button
-                  onClick={() => {
-                    addExternalOrganizerToAssembly(
-                      group.groupSlug,
-                      ntnui_no_of_new_organizer,
-                      "Name"
-                    ).then(() => fetchOrganizers());
-                  }}
-                  m={10}
-                  data-testid="add-organizer-button"
-                >
-                  Add organizer
-                </Button>
-              </Modal>
-              Add organizer
-            </Button>
+            <AddOrganizerButtonModal groupSlug={group.groupSlug} />
 
             <Button
               onClick={() => addCase()}
